completed the debbuging process

+++ app/controllers/toys_controller.rb
@@ -7,19 +7,27 @@ class ToysController < ApplicationController
   end
 
   def create
-    toy = Toys.create(toy_params)
+    #fix the typo in the toy 
+    toy = Toy.create(toy_params)
     render json: toy, status: :created
   end
 
   def update
     toy = Toy.find_by(id: params[:id])
     toy.update(toy_params)
+    #render the json 
+    render json: toy 
   end
 
   def destroy
     toy = Toy.find_by(id: params[:id])
-    toy.destroy
-    head :no_content
+    if toy 
+      toy.destroy
+      head :no_content
+    else
+      render json:{error: "Toy not found"}
+    end
+    
   end
 
   private
